Module google_libraryagent1::api

Structs

GoogleExampleLibraryagentV1Book
A single book in the library.

GoogleExampleLibraryagentV1ListBooksResponse
Response message for LibraryAgent.ListBooks.

GoogleExampleLibraryagentV1ListShelvesResponse
Response message for LibraryAgent.ListShelves.

GoogleExampleLibraryagentV1Shelf
A Shelf contains a collection of books with a theme.

Libraryagent
Central instance to access all Libraryagent related resource activities

ShelveBookBorrowCall
Borrow a book from the library. Returns the book if it is borrowed successfully. Returns NOT_FOUND if the book does not exist in the library. Returns quota exceeded error if the amount of books borrowed exceeds allocation quota in any dimensions.

ShelveBookGetCall
Gets a book. Returns NOT_FOUND if the book does not exist.

ShelveBookListCall
Lists books in a shelf. The order is unspecified but deterministic. Newly created books will not necessarily be added to the end of this list. Returns NOT_FOUND if the shelf does not exist.

ShelveBookReturnCall
Return a book to the library. Returns the book if it is returned to the library successfully. Returns error if the book does not belong to the library or the users didn't borrow before.

ShelveGetCall
Gets a shelf. Returns NOT_FOUND if the shelf does not exist.

ShelveListCall
Lists shelves. The order is unspecified but deterministic. Newly created shelves will not necessarily be added to the end of this list.

ShelveMethods
A builder providing access to all methods supported on shelve resources. It is not used directly, but through the Libraryagent hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.